+++ finite-volume/source/operators/interpolation_pc.c
@@ -26,7 +26,7 @@ void cb_copy_int(double *buf, int n, int srcid, int depth_f, int id_f, int pcl, 
   else prescale_f = level_f->prescale_fl;
 
   int i;
-  int nth = MAX_TLVG*level_f->my_rank + MAX_LVG*6 + MAX_VG*level_f->depth + MAX_NBGS*(id_f*2+pcl);
+  size_t nth = MAX_TLVG*(size_t)level_f->my_rank + MAX_LVG*6 + MAX_VG*level_f->depth + MAX_NBGS*(id_f*2+pcl);
   for (i = 0; i < level_f->interpolation.num_recvs; i++) {
      if (level_f->interpolation.recv_ranks[i] == srcid) {
         if (upc_rflag[nth+i] != 0) {
@@ -231,7 +231,7 @@ void interpolation_pc(level_type * level_f, int id_f, double prescale_f, level_t
     } else {
       int rid = level_c->interpolation.send_ranks[n];
       int pos = level_c->interpolation.send_match_pos[n];
-      int nth = MAX_TLVG*rid + MAX_LVG*6 + MAX_VG*level_f->depth + MAX_NBGS*(id_f*2);
+      size_t nth = MAX_TLVG*(size_t)rid + MAX_LVG*6 + MAX_VG*level_f->depth + MAX_NBGS*(id_f*2);
       upc_rflag[nth+pos] = 1;
       nshm++;
     }
@@ -256,7 +256,7 @@ void interpolation_pc(level_type * level_f, int id_f, double prescale_f, level_t
 #ifdef USE_UPCXX
 #ifdef UPCXX_AM
 
-  int nth = MAX_TLVG*level_f->my_rank + MAX_LVG*6 + MAX_VG*level_f->depth + MAX_NBGS*id_f*2;
+  size_t nth = MAX_TLVG*(size_t)level_f->my_rank + MAX_LVG*6 + MAX_VG*level_f->depth + MAX_NBGS*id_f*2;
   while (1) {
     int arrived = 0;
     for (int n = 0; n < level_f->interpolation.num_recvs; n++) {
